requirements.txt
# Django Core
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Task Queue
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.6.0
redis==5.0.3

# Mercado Pago
mercadopago==2.2.3

# Validation
validate-docbr==1.11.1

# QR Code
qrcode==8.0
Pillow==10.2.0

# Utils
python-dotenv==1.0.1
dj-static==0.0.6
gunicorn==21.2.0
django-filter==24.1
django-cors-headers==4.3.1

# Email
django-celery-email==3.0.0

# Documentation
drf-spectacular==0.27.1
